% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_genes}
\alias{append_genes}
\title{Given a reference matrix and a list of genes, take the union of
all genes in vector and genes in reference matrix
and insert zero counts for all remaining genes.}
\usage{
append_genes(gene_vector, ref_matrix)
}
\arguments{
\item{ref_matrix}{Reference matrix containing cell types vs.
gene expression values}

\item{geneVector}{char vector with gene names}
}
\value{
Reference matrix with union of all genes
}
\description{
Given a reference matrix and a list of genes, take the union of
all genes in vector and genes in reference matrix
and insert zero counts for all remaining genes.
}
