% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_ident}
\alias{assign_ident}
\title{manually change idents as needed}
\usage{
assign_ident(metadata, cluster_col = "cluster", ident_col = "type",
  clusters, idents)
}
\arguments{
\item{metadata}{column of ident}

\item{cluster_col}{column in metadata containing cluster info}

\item{ident_col}{column in metadata containing identity assignment}

\item{clusters}{names of clusters to change, string or vector of strings}

\item{idents}{new idents to assign, must be length of 1 or same as clusters}
}
\value{
new dataframe of metadata
}
\description{
manually change idents as needed
}
