% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_clusters}
\alias{average_clusters}
\title{Average expression values per cluster}
\usage{
average_clusters(mat, cluster_info, log_scale = T, cell_col = "rn",
  cluster_col = "cluster")
}
\arguments{
\item{mat}{expression matrix}

\item{cluster_info}{data.frame with cells}

\item{log_scale}{input data is natural log,
averaging will be done on unlogged data}

\item{cell_col}{column in cluster_info with cell ids}

\item{cluster_col}{column in cluster_info with cluster number}
}
\description{
Average expression values per cluster
}
