% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_clusters_filter}
\alias{average_clusters_filter}
\title{Average expression values per cluster, filtered by set parameter, defaults to calculating background}
\usage{
average_clusters_filter(
  mat,
  metadata,
  if_log = TRUE,
  filter_on = "nGene",
  group_by = NULL,
  filter_method = "<=",
  filter_value = 300
)
}
\arguments{
\item{mat}{expression matrix}

\item{metadata}{data.frame or vector containing cluster assignments per cell, and attribute to filter on.
Order must match column order in supplied matrix. If a data.frame
provide the cluster_col parameters.}

\item{if_log}{input data is natural log,
averaging will be done on unlogged data}

\item{filter_on}{column in metadata to filter on}

\item{group_by}{column name to use for cluster identity}

\item{filter_method}{"<", "==", ">" compared to filter_value}

\item{filter_value}{baseline minimum as background cutoff}
}
\value{
average expression matrix, with genes for row names, and clusters for column names
}
\description{
Average expression values per cluster, filtered by set parameter, defaults to calculating background
}
\examples{
average_clusters_filter(
    mat = pbmc_matrix_small,
    metadata = pbmc_meta,
    group_by = "classified",
    filter_on = "seurat_clusters",
    filter_method = "==",
    filter_value = "1"
)
}
