% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_seurat_meta}
\alias{build_seurat_meta}
\title{Build meta-data for Seurat object}
\usage{
build_seurat_meta(x, metadata, cluster_col, threshold, per_cell,
  rename_suff)
}
\description{
Build meta-data for Seurat object
}
