% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_pathway_gsea}
\alias{calculate_pathway_gsea}
\title{Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)}
\usage{
calculate_pathway_gsea(mat, pathway_list, n_perm = 1000, scale = T,
  no_warnings = T)
}
\arguments{
\item{mat}{expression matrix}

\item{pathway_list}{a list of vectors, each named for a specific pathway, or dataframe}

\item{n_perm}{Number of permutation for fgsea function. Defaults to 1000.}

\item{scale}{convert expr_mat into zscores prior to running GSEA?, default = FALSE}
}
\description{
Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)
}
