% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dplyr.R
\name{call_consensus}
\alias{call_consensus}
\title{get concensus calls for a list of cor calls}
\usage{
call_consensus(list_of_res)
}
\arguments{
\item{list_of_res}{list of call dataframes from cor_to_call_rank}
}
\value{
dataframe of cluster, new ident, and mean rank
}
\description{
get concensus calls for a list of cor calls
}
\examples{
res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  cluster_col = "classified",
  ref_mat = pbmc_bulk_matrix
)

res2 <- cor_to_call_rank(res, threshold = "auto")
res3 <- cor_to_call_rank(res)
res4 <- call_consensus(list(res2, res3))
}
