% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbmc_m}
\alias{cbmc_m}
\title{reference marker matrix from seurat citeseq CBMC tutorial}
\format{An object of class \code{data.frame} with 3 rows and 13 columns.}
\source{
(https://satijalab.org/seurat/v3.0/multimodal_vignette.html#identify-differentially-expressed-proteins-between-clusters)
}
\usage{
cbmc_m
}
\description{
reference marker matrix from seurat citeseq CBMC tutorial
}
\seealso{
Other data: 
\code{\link{cbmc_ref}},
\code{\link{downrefs}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}},
\code{\link{s_small3}},
\code{\link{s_small}}
}
\concept{data}
\keyword{datasets}
