% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbmc_ref}
\alias{cbmc_ref}
\title{reference matrix from seurat citeseq CBMC tutorial}
\format{expression matrix
\describe{
}}
\source{
(https://satijalab.org/seurat/v3.0/multimodal_vignette.html#identify-differentially-expressed-proteins-between-clusters)
}
\usage{
cbmc_ref
}
\description{
reference matrix from seurat citeseq CBMC tutorial
}
\keyword{datasets}
