% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_raw_counts}
\alias{check_raw_counts}
\title{Given a count matrix, determine if the matrix has been either
log-normalized, normalized, or contains raw counts}
\usage{
check_raw_counts(counts_matrix, max_log_value = 50)
}
\arguments{
\item{counts_matrix}{Count matrix containing scRNA-seq read data}

\item{max_log_value}{Static value to determine if a matrix is normalized}
}
\value{
String either raw counts, log-normalized or normalized
}
\description{
Given a count matrix, determine if the matrix has been either
log-normalized, normalized, or contains raw counts
}
\examples{
mat <- append_genes(
gene_vector = human_genes_10x,
ref_matrix = cbmc_ref 
) 
}
