% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clustify}
\alias{clustify}
\title{Main function to compare scRNA-seq data to bulk RNA-seq data.}
\usage{
clustify(expr_mat, metadata, bulk_mat, query_genes, cluster_col = NULL,
  cell_col = NULL, per_cell = F, num_perm = 0, return_full = F,
  compute_method = "spearman", ...)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{metadata}{clustering info of single-cell data}

\item{bulk_mat}{bulk expression matrix}

\item{query_genes}{A vector of genes of interest.}

\item{cluster_col}{column in metadata that contains cluster ids per cell. Will default to first
column of metadata if not supplied. Not required if running correlation per cell.}

\item{cell_col}{column in metadata that contains the cell ids that match to columns
in the single-cell expression matrix. Will default to rownames of the metadata data.frame
if not supplied. Not required if running correlation per cluster.}

\item{per_cell}{if true run per cell, otherwise per cluster.}

\item{num_perm}{number of permutations, set to 0 by default}

\item{return_full}{Return full results includeing scores instead
of correlation coefficient only.}

\item{compute_method}{method(s) for computing similarity scores}

\item{...}{additional arguments to pass to compute_method function}
}
\description{
Main function to compare scRNA-seq data to bulk RNA-seq data.
}
