% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clustify_nudge}
\alias{clustify_nudge}
\title{using marker gene percentage info to nudge cor calling}
\usage{
clustify_nudge(input, bulk_mat, marker, cluster_col = NULL,
  query_genes = NULL, compute_method = "spearman", weight = 0.2,
  seurat_out = F, threshold = 0, dr = "tsne", set_ident = T)
}
\arguments{
\item{input}{seurat 2 object}

\item{bulk_mat}{bulk expression matrix}

\item{cluster_col}{column in metadata that contains cluster ids per cell. Will default to first
column of metadata if not supplied. Not required if running correlation per cell.}

\item{query_genes}{A vector of genes of interest to compare. If NULL, then common genes between
the expr_mat and bulk_mat will be used for comparision.}

\item{compute_method}{method(s) for computing similarity scores}

\item{seurat_out}{output cor matrix or called seurat object}

\item{dr}{stored dimension reduction}

\item{...}{additional arguments to pass to compute_method function}
}
\description{
using marker gene percentage info to nudge cor calling
}
