% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{clustify_seurat}
\alias{clustify_seurat}
\title{Wrapper to integate into Seurat workflow}
\usage{
clustify_seurat(seurat_object, bulk_mat, query_gene_list = "var.genes",
  per_cell = F, num_perm = 0, cluster_col = "cluster",
  compute_method = corr_coef, output = "object", carry_cor = FALSE)
}
\arguments{
\item{seurat_object}{seruat_object after tsne projections and clustering}

\item{bulk_mat}{bulk expression matrix, or matrixized ranked list}

\item{query_gene_list}{A vector of genes of interest, or by default loads var.genes from seurat object, a number can be given to subset var.genes}

\item{per_cell}{if true run per cell, otherwise per cluster.}

\item{num_perm}{number of permutations, set to 0 by default}

\item{compute_method}{method(s) for computing similarity scores}

\item{output}{either output a new seurat object, or just a dataframe of calls}

\item{return_full}{Return full results includeing scores instead
of correlation coefficient only.}

\item{...}{additional arguments to pass to compute_method function}
}
\description{
Wrapper to integate into Seurat workflow
}
