% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustifyR-package.R
\docType{package}
\name{clustifyr-package}
\alias{clustifyr-package}
\alias{_PACKAGE}
\title{clustifyr: Classifier for Single-cell RNA-seq Using Cell Clusters}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Classify cells from single-cell RNA sequencing data
    using external reference data (e.g., bulk RNA-seq).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/rnabioco/clustifyr#readme}
  \item \url{https://rnabioco.github.io/clustifyr/}
  \item Report bugs at \url{https://github.com/rnabioco/clustifyr/issues}
}

}
\author{
\strong{Maintainer}: Rui Fu \email{raysinensis@gmail.com}

Authors:
\itemize{
  \item Kent Riemondy \email{kent.riemondy@ucdenver.edu}
  \item Austin Gillen \email{austin.gillen@ucdenver.edu}
  \item Chengzhe Tian \email{Chengzhe.Tian@colorado.edu}
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com}
}

Other contributors:
\itemize{
  \item Yue Hao \email{haoyuethink@gmail.com} [contributor]
  \item Michelle Daya \email{michelle.daya@ucdenver.edu} [contributor]
}

}
\keyword{internal}
