% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustifyR-package.R
\docType{package}
\name{clustifyR}
\alias{clustifyR}
\alias{clustifyR-package}
\title{clustifyR}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Classify cells from single-cell RNA sequencing data using bulk
  RNA-seq reference data.
}
\seealso{
Report bugs at \url{https://github.com/rnabioco/clustifyR/issues}
}
\author{
\strong{Maintainer}: Kent Riemondy \email{kent.riemondy@ucdenver.edu}

Authors:
\itemize{
  \item Austin Gillen \email{austin.gillen@ucdenver.edu}
  \item Rui Fu \email{rui.fu@ucdenver.edu}
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com}
  \item Yue Hao \email{haoyuethink@gmail.com}
  \item Chengzhe Tian \email{Chengzhe.Tian@colorado.edu}
  \item Michelle Daya \email{michelle.daya@ucdenver.edu}
}

}
