% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{clustifyr_methods}
\alias{clustifyr_methods}
\title{Correlation functions available in clustifyr}
\format{An object of class \code{character} of length 5.}
\usage{
clustifyr_methods
}
\description{
Correlation functions available in clustifyr
}
\keyword{datasets}
