% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_mean_expr}
\alias{compute_mean_expr}
\title{compute mean of clusters}
\usage{
compute_mean_expr(expr_mat, sc_assign, sc_clust)
}
\arguments{
\item{expr_mat}{matrix of gene expression}

\item{sc_assign}{vector of cluster assignments}

\item{sc_clust}{unique vector of cluster assignments}
}
\value{
matrix of average expression per cluster
}
\description{
compute mean of clusters
}
