% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Compute similarity between two vectors}
\usage{
compute_similarity(vec1, vec2, compute_method, ...)
}
\arguments{
\item{vec1}{test vector}

\item{vec2}{reference vector}

\item{compute_method}{method to run i.e. corr_coef}
}
\description{
Compute the similarity score between two vectors using a customized scoring function
vec1, vec2: input expression vectors, can come from single cell or bulk RNA seq data.
the length of vec1 and vec2 must match, and corresponding elements must refer to the same gene.
vec1 and vec2 should be AFTER pre-processing, feature selection and dimension reduction
compute_method and ...: function to compute the similarity score, and relevant parameters for compute_method
}
