% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{corr_coef}
\alias{corr_coef}
\title{Correlation function}
\usage{
corr_coef(vec1, vec2, method = "pearson")
}
\arguments{
\item{vec1}{test vector}

\item{vec2}{reference vector}

\item{method}{pearson, spearman, cosine}
}
\description{
candidate scoring methods for similarity.
score must be between -1 and 1.
\enumerate{
\item completely positively correlated
-1. completely negatively correlated
}
}
