% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{cosine}
\alias{cosine}
\title{Cosine distance}
\usage{
cosine(vec1, vec2)
}
\arguments{
\item{vec1}{test vector}

\item{vec2}{reference vector}
}
\value{
numeric value of cosine distance between the vectors
}
\description{
Cosine distance
}
