% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downrefs}
\alias{downrefs}
\title{table of references stored in clustifyrdata}
\format{dataframe
\describe{
}}
\source{
various packages
}
\usage{
downrefs
}
\description{
table of references stored in clustifyrdata
}
\keyword{datasets}
