% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsce_small}
\alias{fsce_small}
\title{Small clustered FSCE object}
\format{XXX
\describe{
}}
\source{
\code{[fsce_small]} processed by \code{[Scrunchy]}
}
\usage{
fsce_small
}
\description{
Small clustered FSCE object
}
\keyword{datasets}
