% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gene_pct}
\alias{gene_pct}
\title{pct of cells in each cluster that express genelist}
\usage{
gene_pct(matrix, genelist, clusters)
}
\arguments{
\item{matrix}{expression matrix}

\item{genelist}{vector of marker genes for one identity}

\item{clusters}{vector of cluster identities}
}
\description{
pct of cells in each cluster that express genelist
}
