% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_best_str}
\alias{get_best_str}
\title{Function to make call and attach score}
\usage{
get_best_str(name, best_mat, cor_mat, carry_cor = TRUE)
}
\arguments{
\item{name}{name of row to query}

\item{best_mat}{binarized call matrix}

\item{cor_mat}{correlation matrix}

\item{carry_cor}{whether the correlation score gets reported}
}
\description{
Function to make call and attach score
}
\examples{
cor_mat <- res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  cluster_col = "classified",
  ref_mat = pbmc_bulk_matrix,
  query_genes = pbmc_vargenes
)

best_mat <- get_best_match_matrix(cor_mat)

get_best_str(
  name = 1,
  best_mat = best_mat,
  cor_mat = cor_mat
)
}
