% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_column}
\alias{get_unique_column}
\title{Generate a unique column id for a dataframe}
\usage{
get_unique_column(df, id = NULL)
}
\arguments{
\item{df}{dataframe with column names}

\item{id}{desired id if unique}
}
\value{
character
}
\description{
Generate a unique column id for a dataframe
}
