% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genelist.R
\name{get_vargenes}
\alias{get_vargenes}
\title{Generate variable gene list from marker matrix}
\usage{
get_vargenes(marker_mat)
}
\arguments{
\item{marker_mat}{matrix or dataframe of candidate genes for each cluster}
}
\value{
vector of marker gene names
}
\description{
Variable gene list is required for \code{clustify} main function. This
function parses variables genes from a matrix input.
}
\examples{
get_vargenes(cbmc_m)
}
