% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_genes_10x}
\alias{human_genes_10x}
\title{Vector of human genes for 10x cellranger pipeline}
\format{
An object of class \code{character} of length 33514.
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/downloads/latest}
}
\usage{
human_genes_10x
}
\description{
Vector of human genes for 10x cellranger pipeline
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{cbmc_ref}},
\code{\link{downrefs}},
\code{\link{mouse_genes_10x}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}},
\code{\link{s_small3}},
\code{\link{s_small}},
\code{\link{sce_small}}
}
\concept{data}
\keyword{datasets}
