% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genelist.R
\name{hyperp}
\alias{hyperp}
\title{calculate adjusted p-values for hypergeometric test of gene lists}
\usage{
hyperp(bin_matrix, marker_m, n = 30000)
}
\arguments{
\item{bin_matrix}{binarized single-cell expression matrix}

\item{marker_m}{matrix or dataframe of candidate genes for each cluster}

\item{n}{number of genes in the genome}
}
\description{
calculate adjusted p-values for hypergeometric test of gene lists
}
