% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_heatmap.R
\name{important_heatmap}
\alias{important_heatmap}
\title{Heatmap of average expression of most predictive genes across cell types}
\usage{
important_heatmap(important, avg_matrix, meta, MDA_thresh, MDG_thresh)
}
\arguments{
\item{important}{importance measurements of genes (predictors) from the random forest model}

\item{avg_matrix}{average expression matrix}

\item{meta}{contains cluster info and classified cell types}

\item{MDA_thresh}{MeanDecreaseAccuracy threshold}

\item{MDG_thresh}{MeanDecreaseGini threshold}
}
\description{
Heatmap of average expression of most predictive genes across cell types
}
