% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_meta_object}
\alias{insert_meta_object}
\title{more flexible metadata update of single cell objects}
\usage{
insert_meta_object(input, new_meta, type = class(input),
  meta_loc = NULL, lookuptable = NULL)
}
\arguments{
\item{input}{input object}

\item{new_meta}{new metadata table to insert back into object}

\item{type}{look up predefined slots/loc}

\item{meta_loc}{metadata location}

\item{lookuptable}{if not supplied, will look in built-in table for object parsing}
}
\value{
new object with new metadata inserted
}
\description{
more flexible metadata update of single cell objects
}
\examples{
\dontrun{
clustifyr_obj <- insert_meta_object(s_small3, seurat_meta(s_small3, dr = "tsne"))
}
}
