% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pkg_available}
\alias{is_pkg_available}
\title{Check package is installed}
\usage{
is_pkg_available(pkg, action = c("none", "message", "warn", "error"), msg = "")
}
\arguments{
\item{pkg}{package to query}
}
\value{
logical(1) indicating if package is available.
}
\description{
Check package is installed
}
