% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_comb_ref}
\alias{make_comb_ref}
\title{make combination ref matrix to assess intermixing}
\usage{
make_comb_ref(ref_mat, log_scale = T, sep = "_and_")
}
\arguments{
\item{ref_mat}{reference expression matrix}

\item{log_scale}{whether input data is natural}

\item{sep}{separator for name combinations}
}
\description{
make combination ref matrix to assess intermixing
}
