% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{not_pretty_palette}
\alias{not_pretty_palette}
\title{black and white palette for plotting continous variables}
\format{An object of class \code{character} of length 9.}
\usage{
not_pretty_palette
}
\value{
vector of colors
}
\description{
black and white palette for plotting continous variables
}
\examples{
clustifyr:::not_pretty_palette
}
\keyword{datasets}
