% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_loc_lookup}
\alias{object_loc_lookup}
\title{lookup table for single cell object structures}
\usage{
object_loc_lookup()
}
\value{
A list populated with standardized functions to
access relevant data structures in multiple single cell
data formats.
}
\description{
lookup table for single cell object structures
}
