% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overcluster}
\alias{overcluster}
\title{Overcluster by kmeans per cluster}
\usage{
overcluster(mat, cluster_id, power = 0.15, seed = 42)
}
\arguments{
\item{mat}{expression matrix}

\item{cluster_id}{list of ids per cluster}

\item{power}{decides the number of clusters for kmeans}

\item{seed}{seed for kmeans}
}
\value{
new cluster_id list of more clusters
}
\description{
Overcluster by kmeans per cluster
}
\examples{
pbmc_ids <- overcluster(
  mat = pbmc_matrix_small,
  cluster_id = split(colnames(pbmc_matrix_small), pbmc_meta$classified)
)
}
