% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overcluster_test}
\alias{overcluster_test}
\title{compare clustering parameters and classification outcomes}
\usage{
overcluster_test(expr, metadata, cluster_col, x_col = "tSNE_1",
  y_col = "tSNE_2", n = 5, do.label = T, seed = 42,
  newclustering = NULL)
}
\arguments{
\item{expr}{expression matrix}

\item{metadata}{metadata including cluster info and dimension reduction plotting}

\item{cluster_col}{column of clustering from metadata}

\item{x_col}{column of metadata for x axis plotting}

\item{y_col}{column of metadata for y axis plotting}

\item{n}{expand n-fold for over/under clustering}

\item{do.label}{whether to label each cluster at median center}

\item{seed}{set seed for kmeans}

\item{newclustering}{use kmeans if NULL on dr or col name for second column of clustering}
}
\description{
compare clustering parameters and classification outcomes
}
