% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_loc_object}
\alias{parse_loc_object}
\title{more flexible parsing of single cell objects}
\usage{
parse_loc_object(
  input,
  type = class(input),
  expr_loc = NULL,
  meta_loc = NULL,
  var_loc = NULL,
  cluster_col = NULL,
  lookuptable = NULL
)
}
\arguments{
\item{input}{input object}

\item{type}{look up predefined slots/loc}

\item{expr_loc}{function that extracts expression matrix}

\item{meta_loc}{function that extracts metadata}

\item{var_loc}{function that extracts variable genes}

\item{cluster_col}{column of clustering from metadata}

\item{lookuptable}{if not supplied, will use object_loc_lookup() for parsing.}
}
\value{
list of expression, metadata, vargenes, cluster_col info from object
}
\description{
more flexible parsing of single cell objects
}
\examples{
so <- so_pbmc() 
obj <- parse_loc_object(so)
length(obj)
}
