% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_avg}
\alias{pbmc4k_avg}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq PBMCs.}
\format{A sparseMatrix with genes as rows and clusters as columns.}
\source{
\code{[pbmc4k_matrix]} averaged per cluster
}
\usage{
pbmc4k_avg
}
\description{
A sample of 4000 PBMC cells from a bigger experiment
}
\keyword{datasets}
