% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_markers}
\alias{pbmc4k_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{A data frame with 7 variables:}
\source{
\code{[pbmc4k_matrix]} processed by \code{[seurat]}
}
\usage{
pbmc4k_markers
}
\description{
A sample of 4000 cells from a bigger experiment
}
\keyword{datasets}
