% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_meta}
\alias{pbmc4k_meta}
\title{Meta-data for single-cell RNA-seq PBMCs.}
\format{A data frame with 5 variables:
\describe{
}}
\source{
\code{[pbmc4k_matrix]} processed by \code{[seurat]}
}
\usage{
pbmc4k_meta
}
\description{
A sample of 300 PBMC cells, downsampled from a bigger experiment, 30 cells per cluster
}
\keyword{datasets}
