% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_vargenes}
\alias{pbmc4k_vargenes}
\title{variable genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{A vector:}
\source{
\code{[pbmc4k_matrix]} processed by \code{[seurat]}
}
\usage{
pbmc4k_vargenes
}
\description{
A sample of 4000 cells from a bigger experiment
}
\keyword{datasets}
