% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc5_matrix}
\alias{pbmc5_matrix}
\title{Matrix of single-cell RNA-seq PBMCs from 5'end kit.}
\format{A data frame with 13 variables:
\describe{
}}
\source{
\link{google.com}
}
\usage{
pbmc5_matrix
}
\description{
A sample of 8000 cells from a bigger experiment
}
\keyword{datasets}
