% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc5_meta}
\alias{pbmc5_meta}
\title{Meta-data for single-cell RNA-seq PBMCs from 5'end kit.}
\format{A data frame with 4 variables:
\describe{
}}
\source{
\link{google.com}
}
\usage{
pbmc5_meta
}
\description{
A sample of 8000 cells from a bigger experiment
}
\keyword{datasets}
