% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_bulk_matrix}
\alias{pbmc_bulk_matrix}
\title{Bulk RNA-Seq data from sorted populations isolated from PBMCs.}
\format{A read count matrix with 14 variables:}
\source{
(http://duffel.rail.bio/recount/v2/SRP051688/rse_gene.Rdata)
}
\usage{
pbmc_bulk_matrix
}
\description{
Bulk RNA-Seq data from sorted populations isolated from PBMCs.
}
\keyword{datasets}
