% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers}
\alias{pbmc_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{A data frame with 7 variables:
\describe{
}}
\source{
\code{[pbmc_matrix]} processed by \code{[Seurat]}
}
\usage{
pbmc_markers
}
\description{
Dataframe of markers from Seurat FindAllMarkers function
}
\keyword{datasets}
