% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers_M3Drop}
\alias{pbmc_markers_M3Drop}
\title{Marker genes identified by M3Drop from single-cell RNA-seq PBMCs.}
\format{A data frame with 3 variables:
\describe{
}}
\source{
\code{[pbmc_matrix]} processed by \code{[M3Drop]}
}
\usage{
pbmc_markers_M3Drop
}
\description{
Selected features of 3k pbmcs from Seurat3 tutorial
}
\keyword{datasets}
