% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_matrix_small}
\alias{pbmc_matrix_small}
\title{Matrix of single-cell RNA-seq PBMCs.}
\format{A sparseMatrix with genes as rows and cells as columns.

\describe{
}
R}
\source{
(https://satijalab.org/seurat/v3.0/pbmc3k_tutorial.html)
}
\usage{
pbmc_matrix_small
}
\description{
Count matrix of 3k pbmcs from Seurat3 tutorial, with only var.features
}
\keyword{datasets}
