% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_meta}
\alias{pbmc_meta}
\title{Meta-data for single-cell RNA-seq PBMCs.}
\format{A data frame with 5 variables:
\describe{
}}
\source{
\code{[pbmc_matrix]} processed by Seurat
}
\usage{
pbmc_meta
}
\description{
Metadata, including umap, of 3k pbmcs from Seurat3 tutorial
}
\keyword{datasets}
