% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_pca}
\alias{pbmc_pca}
\title{Variable genes defined by Principal components analysis of pbmc4k data}
\format{XXX
\describe{
}}
\source{
\code{[pbmc4k_matrix]} processed by \code{[seurat]}
}
\usage{
pbmc_pca
}
\description{
Variable genes defined by Principal components analysis of pbmc4k data
}
\keyword{datasets}
