% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_vargenes}
\alias{pbmc_vargenes}
\title{Variable genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{A vector:
\describe{
}}
\source{
\verb{[pbmc_matrix]} processed by Seurat
}
\usage{
pbmc_vargenes
}
\description{
Top 2000 variable genes from 3k pbmcs from Seurat3 tutorial
}
\keyword{datasets}
