% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_similarity.R
\name{permutation_similarity}
\alias{permutation_similarity}
\title{compute the p-value for similarity using permutation}
\usage{
permutation_similarity(sc_expr, bulk_expr, sc_meta, num_perm, compute_method,
  metadata = "cluster", ...)
}
\arguments{
\item{compute_method}{parameters feed in for computing similarity score}

\item{metadata}{metadata column used for clustering}

\item{sc_expr:}{single-cell expr matrix}

\item{bulk_expr:}{bulk epxr matrix.}

\item{sc_meta:}{clustering info of single-cell data assume that genes have ALREADY BEEN filtered}

\item{num_perm:}{number of permutation}
}
\description{
Permute cluster labels to calculate empirical p-value
}
