% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{permute_similarity}
\alias{permute_similarity}
\title{Compute a p-value for similarity using permutation}
\usage{
permute_similarity(expr_mat, ref_mat, cluster_ids, n_perm,
  per_cell = FALSE, compute_method, rm0 = FALSE, ...)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{ref_mat}{reference expression matrix}

\item{cluster_ids}{clustering info of single-cell data assume that
genes have ALREADY BEEN filtered}

\item{n_perm}{number of permutations}

\item{per_cell}{run per cell?}

\item{compute_method}{method(s) for computing similarity scores}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{...}{additional parameters}
}
\value{
matrix of numeric values
}
\description{
Permute cluster labels to calculate empirical p-value
}
\examples{
permute_similarity(
  expr_mat = pbmc_matrix_small[c("PPBP", "LYZ", "S100A9"), c(7, 11)],
  ref_mat = cbmc_ref[c("PPBP", "LYZ", "S100A9"), 1:3],
  cluster_ids = colnames(pbmc_matrix_small[c("PPBP", "LYZ", "S100A9"), c(7, 11)]),
  n_perm = 2,
  per_cell = TRUE,
  compute_method = "spearman"
)
}
