% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_call}
\alias{plot_call}
\title{Plot called clusters on a tSNE, for each reference cluster given}
\usage{
plot_call(cor_matrix, metadata, data_to_plot = colnames(cor_matrix), ...)
}
\arguments{
\item{cor_matrix}{input similarity matrix}

\item{metadata}{input metadata with tsne coordinates and cluster ids}

\item{data_to_plot}{colname of data to plot, defaults to all}

\item{...}{passed to plot_tsne}
}
\description{
Plot called clusters on a tSNE, for each reference cluster given
}
