% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cols}
\alias{plot_cols}
\title{Plot variable median per cluster from reference metadata vs new assigned metadata, for visually evaluating classification}
\usage{
plot_cols(metadata, cluster_col, cluster_col_called, plot_col,
  metadata_ref, cluster_col_ref, plot_col_ref)
}
\arguments{
\item{metadata}{clustify-ed metadata}

\item{cluster_col}{metadata column for original clustering}

\item{cluster_col_called}{metadata column for clustifyr assignments}

\item{plot_col}{metadata column to plot}

\item{metadata_ref}{reference single cell RNA seq metadata}

\item{cluster_col_ref}{metadata column for cluster for reference}

\item{plot_col_ref}{metadata column to plot for reference}
}
\value{
ggplot scatterplot
}
\description{
Plot variable median per cluster from reference metadata vs new assigned metadata, for visually evaluating classification
}
