% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot similarity measures on a tSNE}
\usage{
plot_cor(correlation_matrix, metadata,
  bulk_data_to_plot = colnames(correlation_matrix), cluster_col = NULL,
  dim1_col = "tSNE_1", dim2_col = "tSNE_2", scale_legends = FALSE)
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{metadata}{input metadata with per cell tsne coordinates and cluster ids}

\item{bulk_data_to_plot}{colname of data to plot, defaults to all}

\item{cluster_col}{colname of clustering data in metadata, defaults to rownames of the
metadata if not supplied.}

\item{dim1_col}{metadata column name with 1st axis dimension.
defaults to "tSNE_1".}

\item{dim2_col}{metadata column name with 2nd axis dimension.
defaults to "tSNE_2".}

\item{scale_legends}{if TRUE scale all legends to maximum values in entire
correlation matrix. if FALSE scale legends to maximum for each plot. A
two-element numeric vector can also be passed to supply custom values i.e. c(0, 1)}

\item{...}{passed to plot_tsne}
}
\description{
Plot similarity measures on a tSNE
}
