% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{Plot similarity measures on heatmap}
\usage{
plot_cor_heatmap(correlation_matrix, metadata = NULL,
  cluster_col = NULL, col = not_pretty_palette, ...)
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{metadata}{input metadata with per cell tsne coordinates and cluster ids}

\item{cluster_col}{colname of clustering data in metadata, defaults to rownames of the
metadata if not supplied.}

\item{col}{color ramp to use}

\item{...}{passed to Heatmap}
}
\description{
Plot similarity measures on heatmap
}
