% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot gene expression on to tSNE}
\usage{
plot_gene(expr_mat, metadata, genes, cell_col = NULL, ...)
}
\arguments{
\item{expr_mat}{input single cell matrix}

\item{metadata}{data.frame with tSNE coordinates}

\item{genes}{gene(s) to color tSNE}

\item{cell_col}{column name in metadata containing cell ids, defaults
to rownames if not supplied}

\item{...}{additional arguments passed to \code{[clustifyR::plot_tsne()]}}
}
\description{
Plot gene expression on to tSNE
}
