% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pos_neg_marker}
\alias{pos_neg_marker}
\title{generate pos and negative marker expression matrix from a
list/dataframe of positive markers}
\usage{
pos_neg_marker(mat)
}
\arguments{
\item{mat}{matrix or dataframe of markers}
}
\value{
matrix of gene expression
}
\description{
generate pos and negative marker expression matrix from a
list/dataframe of positive markers
}
\examples{
m1 <- pos_neg_marker(cbmc_m)
}
