% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{pretty_palette}
\alias{pretty_palette}
\title{Color palette for plotting continous variables}
\format{An object of class \code{character} of length 6.}
\usage{
pretty_palette
}
\value{
vector of colors
}
\description{
Color palette for plotting continous variables
}
\keyword{datasets}
