% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{pretty_palette2}
\alias{pretty_palette2}
\title{Color palette for plotting continous variables, starting at gray}
\format{An object of class \code{character} of length 9.}
\usage{
pretty_palette2
}
\value{
vector of colors
}
\description{
Color palette for plotting continous variables, starting at gray
}
\examples{
pretty_palette2
}
\keyword{datasets}
