% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pretty_palette_ramp_d}
\alias{pretty_palette_ramp_d}
\title{Expanded color palette ramp for plotting discrete variables}
\usage{
pretty_palette_ramp_d(n)
}
\arguments{
\item{n}{number of colors to use}
}
\description{
Expanded color palette ramp for plotting discrete variables
}
