% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RClusterCT.R
\docType{package}
\name{RClusterCT}
\alias{RClusterCT}
\alias{RClusterCT-package}
\alias{RClusterCT-package}
\title{RClusterCT}
\description{
RClusterCT
}
\seealso{
Report bugs at \url{https://github.com/NCBI-Hackathons/RClusterCT/issues}
}
\author{
\strong{Maintainer}: Kent Riemondy \email{blah@gmail.com}

Authors:
\itemize{
  \item Austin Gillen \email{blah@gmail.com}
  \item Rui Fu \email{blah@gmail.com}
  \item Jay Hesselberth \email{blah@gmail.com}
  \item Yue Hao \email{blah@gmail.com}
  \item Chengzhe Tian \email{blah@gmail.com}
  \item M. Day \email{blah@gmail.com}
}

}
