% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_matrix.R
\name{reduce_expr_matrix}
\alias{reduce_expr_matrix}
\title{Reduce expression matrix to variable genes and binarize.}
\usage{
reduce_expr_matrix(expr_mat, metadata, vargene)
}
\arguments{
\item{expr_mat}{expression expr_matrix with row names as the gene
names and each column for different cell}

\item{metadata}{cluster and cell type info for each cell}

\item{vargene}{a list of gene names for highly variable genes}
}
\description{
Reduce the full expression expr_matrix to only the
highly variable genes,
change the readcounts to binary factors,
prepare the input data for random forest
}
