% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ref_feature_select}
\alias{ref_feature_select}
\title{feature select from reference matrix}
\usage{
ref_feature_select(mat, n = 3000)
}
\arguments{
\item{mat}{reference matrix}

\item{n}{number of genes to return}
}
\description{
feature select from reference matrix
}
