% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ref_marker_select}
\alias{ref_marker_select}
\title{marker selection from reference matrix}
\usage{
ref_marker_select(mat, cut = 0.5, arrange = T, compto = 1)
}
\arguments{
\item{mat}{reference matrix}

\item{cut}{an expression minimum cutoff}

\item{arrange}{whether to arrange (lower means better)}

\item{compto}{compare max expression to the value of next 1 or more}
}
\description{
marker selection from reference matrix
}
