% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_marker_matrix}
\alias{reverse_marker_matrix}
\title{generate negative markers from a list of exclusive positive markers}
\usage{
reverse_marker_matrix(mat)
}
\arguments{
\item{mat}{matrix or dataframe of markers}
}
\description{
generate negative markers from a list of exclusive positive markers
}
