% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowVar}
\alias{RowVar}
\title{get var per row for matrix}
\usage{
RowVar(x, na.rm = TRUE)
}
\arguments{
\item{x}{expression matrix}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations?}
}
\value{
vector of numeric values
}
\description{
get var per row for matrix
}
\examples{
pbmc_small_rowvar <- RowVar(as.matrix(pbmc_matrix_small))
}
