% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_clustifyr_app}
\alias{run_clustifyr_app}
\title{Launch Shiny app version of clustifyr,
may need to run install_clustifyr_app() at first time to install packages}
\usage{
run_clustifyr_app()
}
\value{
instance of shiny app
}
\description{
Launch Shiny app version of clustifyr,
may need to run install_clustifyr_app() at first time to install packages
}
\examples{
\dontrun{
run_clustifyr_app()
}
}
