% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_cor}
\alias{run_cor}
\title{run_cor: Main function to compare scRNA-seq data to bulk RNA-seq data.}
\usage{
run_cor(expr_mat, metadata, bulk_mat, query_gene_list, per_cell = F,
  num_perm = 0, return_full = F, compute_method, ...)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{metadata}{clustering info of single-cell data}

\item{bulk_mat}{bulk expression matrix}

\item{query_gene_list}{A vector of genes of interest.}

\item{per_cell}{run per cell?}

\item{num_perm}{number of permutations, set to 0 by default}

\item{return_full}{Return full results includeing scores instead of correlation coefficient only.}

\item{compute_method}{method(s) for computing similarity scores}

\item{...}{additional arguments to pass to compute_method function}

\item{cluster_col}{column used for clustering}
}
\description{
run_cor: Main function to compare scRNA-seq data to bulk RNA-seq data.
}
