% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_access.R
\name{sce_pbmc}
\alias{sce_pbmc}
\title{An example SingleCellExperiment object}
\usage{
sce_pbmc()
}
\value{
a SingleCellExperiment object populated with data
from the \link{pbmc_matrix_small} scRNA-seq dataset, additionally
annotated with cluster assignments.
}
\description{
An example SingleCellExperiment object
}
