% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_small}
\alias{sce_small}
\title{Small clustered SCE object}
\format{XXX
\describe{
}}
\source{
\code{[sce_small]} subsampled from pancreas data
}
\usage{
sce_small
}
\description{
Small clustered SCE object
}
\keyword{datasets}
