% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{seurat_ref}
\alias{seurat_ref}
\alias{seurat_ref.Seurat}
\title{Function to convert labelled seurat object to avg expression matrix}
\usage{
seurat_ref(seurat_object, ...)

\method{seurat_ref}{Seurat}(
  seurat_object,
  cluster_col = "classified",
  var_genes_only = FALSE,
  assay_name = NULL,
  method = "mean",
  subclusterpower = 0,
  if_log = TRUE,
  ...
)
}
\value{
reference expression matrix, with genes as row names,
and cell types as column names
}
\description{
Function to convert labelled seurat object to avg expression matrix
}
