% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_access.R
\name{so_pbmc}
\alias{so_pbmc}
\title{An example Seurat object}
\usage{
so_pbmc()
}
\value{
a SingleCellExperiment object populated with data
from the \link{pbmc_matrix_small} scRNA-seq dataset, additionally
annotated with cluster assignments.
}
\description{
An example Seurat object
}
