% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{use_seurat3_comp}
\alias{use_seurat3_comp}
\title{Function to convert labelled seurat3 object to avg expression matrix}
\usage{
use_seurat3_comp(seurat_object, cluster_col = "classified",
  var_genes_only = FALSE, assay_name = NULL, method = "mean")
}
\arguments{
\item{seurat_object}{seurat_object after tsne or umap projections and clustering}

\item{cluster_col}{column name where classified cluster names are stored in seurat meta data, cannot be "rn"}

\item{var_genes_only}{whether to keep only var_genes in the final matrix output, could also look up genes used for PCA}

\item{assay_name}{any additional assay data, such as ADT, to include. If more than 1, pass a vector of names}

\item{method}{whether to take mean (default) or median}
}
\description{
Function to convert labelled seurat3 object to avg expression matrix
}
