% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{vector_similarity}
\alias{vector_similarity}
\title{Compute similarity between two vectors}
\usage{
vector_similarity(vec1, vec2, compute_method, ...)
}
\arguments{
\item{vec1}{test vector}

\item{vec2}{reference vector}

\item{compute_method}{method to run i.e. corr_coef}

\item{...}{arguments to pass to compute_method function}
}
\value{
numeric value of desired correlation or distance measurement
}
\description{
Compute the similarity score between two vectors using a
customized scoring function
Two vectors may be from either scRNA-seq or bulk RNA-seq data.
The lengths of vec1 and vec2 must match, and must be arranged in the
same order of genes.
Both vectors should be provided to this function after pre-processing,
feature selection and dimension reduction.
}
