% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_clusters}
\alias{average_clusters}
\title{Average expression values per cluster}
\usage{
average_clusters(mat, cluster_info, log_scale = T,
  cluster_col = "cluster", cell_col = NULL, low_threshold = 0,
  method = "mean", output_log = T)
}
\arguments{
\item{mat}{expression matrix}

\item{cluster_info}{data.frame or vector containing cluster assignments per cell.
Order must match column order in supplied matrix. If a data.frame
provide the cluster_col parameters.}

\item{log_scale}{input data is natural log,
averaging will be done on unlogged data}

\item{cluster_col}{column in cluster_info with cluster number}

\item{cell_col}{if provided, will reorder matrix first}

\item{low_threshold}{option to remove clusters with too few cells}

\item{method}{whether to take mean (default) or median}

\item{output_log}{whether to report log results}
}
\description{
Average expression values per cluster
}
