% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genelist.R
\name{binarize_expr}
\alias{binarize_expr}
\title{Binarize scRNA seq data}
\usage{
binarize_expr(expr_mat, n = 1000, cut = 0)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{n}{number of top expressing genes to keep}

\item{cut}{cut off to set to 0}
}
\description{
Binarize scRNA seq data
}
