% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{calc_similarity}
\alias{calc_similarity}
\title{compute similarity}
\usage{
calc_similarity(query_mat, ref_mat, compute_method, rm0 = FALSE, ...)
}
\arguments{
\item{query_mat}{query data matrix}

\item{ref_mat}{reference data matrix}

\item{compute_method}{method(s) for computing similarity scores}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{...}{additional parameters}
}
\description{
compute similarity
}
