% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbmc_m}
\alias{cbmc_m}
\title{reference marker matrix from seurat citeseq CBMC tutorial}
\format{matrix of markers
\describe{
}}
\source{
(https://satijalab.org/seurat/v3.0/multimodal_vignette.html#identify-differentially-expressed-proteins-between-clusters)
}
\usage{
cbmc_m
}
\description{
reference marker matrix from seurat citeseq CBMC tutorial
}
\keyword{datasets}
