% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbmc_m}
\alias{cbmc_m}
\title{reference marker matrix from seurat citeseq CBMC}
\format{XXX
\describe{
}}
\source{
\code{[seurat]} tutorial, matrixized top markers
}
\usage{
cbmc_m
}
\description{
reference marker matrix from seurat citeseq CBMC
}
\keyword{datasets}
