% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbmc_ref}
\alias{cbmc_ref}
\title{reference matrix from seurat citeseq CBMC}
\format{XXX
\describe{
}}
\source{
\code{[seurat]} tutorial, averaged per cluster
}
\usage{
cbmc_ref
}
\description{
reference matrix from seurat citeseq CBMC
}
\keyword{datasets}
