% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbmc_ref}
\alias{cbmc_ref}
\title{reference matrix from seurat citeseq CBMC tutorial}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2000 rows and 13 columns.
}
\source{
\url{https://satijalab.org/seurat/v3.0/multimodal_vignette.html#identify-differentially-expressed-proteins-between-clusters}
}
\usage{
cbmc_ref
}
\description{
reference matrix from seurat citeseq CBMC tutorial
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{downrefs}},
\code{\link{human_genes_10x}},
\code{\link{mouse_genes_10x}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}}
}
\concept{data}
\keyword{datasets}
