% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustify.R
\docType{methods}
\name{clustify}
\alias{clustify}
\alias{clustify,matrix-method}
\alias{clustify,Seurat-method}
\title{clustify}
\usage{
clustify(expr_mat, ...)

\S4method{clustify}{matrix}(expr_mat, ref_mat, ref_metadata = NULL,
  ref_columns = NULL, query_metadata = NULL, query_genes = NULL,
  cluster_col = NULL, per_cell = FALSE, num_perm = 0,
  compute_method = "pearson", ...)

\S4method{clustify}{Seurat}(expr_mat, use_var_genes = FALSE, ...)
}
\arguments{
\item{expr_mat}{Single-cell expression matrix or Seurat object}

\item{...}{additional arguments to pass to compute_method function}

\item{ref_mat}{Reference bulk expression matrix}

\item{ref_metadata}{Reference metadata.  If provided, will be used to rename study observations with cell names.}

\item{ref_columns}{A list of columns in the reference metadata to use for translating from study observaton to cell name.  Default: the first and second columns.}

\item{query_metadata}{Query cell cluster assignments, supplied as a vector or data.frame. If
data.frame is supplied then \code{cluster_col} needs to be set. Not required if running correlation per cell.}

\item{query_genes}{A vector of genes of interest to compare. If NULL, then common genes between
the expr_mat and ref_mat will be used for comparision.}

\item{cluster_col}{Column in the query metadata that contains cluster ids per cell. Will default to first
column of metadata if not supplied. Not required if running correlation per cell.}

\item{per_cell}{If true run per cell, otherwise per cluster.  Default: FALSE.}

\item{num_perm}{number of permutations.  Default: 0}

\item{compute_method}{Correlation method(s) to use in computing similarity scores.  Default: 'spearman'}

\item{use_var_genes}{If providing a seurat object, should only the variable genes be used?  Default: FALSE.}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Main function to compare scRNA-seq data to bulk RNA-seq data.
}
