% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustify.R
\name{clustify}
\alias{clustify}
\title{Compare scRNA-seq data to bulk RNA-seq data.}
\usage{
clustify(expr_mat, metadata, bulk_mat, query_genes, compute_method,
  per_cell = FALSE, num_perm = 0, return_full = FALSE, ...)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{metadata}{clustering info of single-cell data}

\item{bulk_mat}{bulk expression matrix}

\item{query_genes}{genes of interest}

\item{compute_method}{method(s) for computing similarity scores}

\item{per_cell}{run per cell, otherwise per cluster.}

\item{num_perm}{number of permutations}

\item{...}{additional arguments to pass to compute_method function}

\item{scores}{include scores in result
correlation coefficient only.}
}
\description{
Compare scRNA-seq data to bulk RNA-seq data.
}
