% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clustify}
\alias{clustify}
\alias{clustify.default}
\alias{clustify.seurat}
\alias{clustify.Seurat}
\title{Main function to compare scRNA-seq data to bulk RNA-seq data.}
\usage{
clustify(input, ...)

\method{clustify}{default}(input, ref_mat, metadata = NULL,
  cluster_col = NULL, query_genes = NULL, per_cell = FALSE,
  n_perm = 0, compute_method = "spearman", verbose = FALSE,
  lookuptable = NULL, rm0 = FALSE, ...)

\method{clustify}{seurat}(input, ref_mat, cluster_col = NULL,
  query_genes = NULL, per_cell = FALSE, n_perm = 0,
  compute_method = "spearman", use_var_genes = TRUE, dr = "tsne",
  seurat_out = TRUE, threshold = 0, verbose = FALSE, rm0 = FALSE,
  rename_suff = NULL, ...)

\method{clustify}{Seurat}(input, ref_mat, cluster_col = NULL,
  query_genes = NULL, per_cell = FALSE, n_perm = 0,
  compute_method = "spearman", use_var_genes = TRUE, dr = "tsne",
  seurat_out = TRUE, threshold = 0, verbose = FALSE, rm0 = FALSE,
  rename_suff = NULL, ...)
}
\arguments{
\item{input}{single-cell expression matrix or Seurat object}

\item{...}{additional arguments to pass to compute_method function}

\item{ref_mat}{reference expression matrix}

\item{metadata}{cell cluster assignments, supplied as a vector or data.frame. If
data.frame is supplied then \code{cluster_col} needs to be set. Not required if running correlation per cell.}

\item{cluster_col}{column in metadata that contains cluster ids per cell. Will default to first
column of metadata if not supplied. Not required if running correlation per cell.}

\item{query_genes}{A vector of genes of interest to compare. If NULL, then common genes between
the expr_mat and ref_mat will be used for comparision.}

\item{per_cell}{if true run per cell, otherwise per cluster.}

\item{n_perm}{number of permutations, set to 0 by default}

\item{compute_method}{method(s) for computing similarity scores}

\item{verbose}{whether to report certain variables chosen}

\item{lookuptable}{if not supplied, will look in built-in table for object parsing}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{use_var_genes}{if providing a seurat object, use the variable genes
(stored in seurat_object@var.genes) as the query_genes.}

\item{dr}{stored dimension reduction}

\item{seurat_out}{output cor matrix or called seurat object}

\item{threshold}{identity calling minimum correlation score threshold}

\item{rename_suff}{suffix to add to type and r column names}

\item{ref_mat}{reference expression matrix}

\item{cluster_col}{column in metadata that contains cluster ids per cell. Will default to first
column of metadata if not supplied. Not required if running correlation per cell.}

\item{query_genes}{A vector of genes of interest to compare. If NULL, then common genes between
the expr_mat and ref_mat will be used for comparision.}

\item{per_cell}{if true run per cell, otherwise per cluster.}

\item{n_perm}{number of permutations, set to 0 by default}

\item{compute_method}{method(s) for computing similarity scores}

\item{use_var_genes}{if providing a seurat object, use the variable genes
(stored in seurat_object@var.genes) as the query_genes.}

\item{dr}{stored dimension reduction}

\item{seurat_out}{output cor matrix or called seurat object}

\item{verbose}{whether to report certain variables chosen}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{...}{additional arguments to pass to compute_method function}

\item{ref_mat}{reference expression matrix}

\item{cluster_col}{column in metadata that contains cluster ids per cell. Will default to first
column of metadata if not supplied. Not required if running correlation per cell.}

\item{query_genes}{A vector of genes of interest to compare. If NULL, then common genes between
the expr_mat and ref_mat will be used for comparision.}

\item{per_cell}{if true run per cell, otherwise per cluster.}

\item{n_perm}{number of permutations, set to 0 by default}

\item{compute_method}{method(s) for computing similarity scores}

\item{use_var_genes}{if providing a seurat object, use the variable genes
(stored in seurat_object@var.genes) as the query_genes.}

\item{dr}{stored dimension reduction}

\item{seurat_out}{output cor matrix or called seurat object}

\item{threshold}{identity calling minimum correlation score threshold}

\item{verbose}{whether to report certain variables chosen}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{rename_suff}{suffix to add to type and r column names}

\item{...}{additional arguments to pass to compute_method function}
}
\description{
Main function to compare scRNA-seq data to bulk RNA-seq data.
}
