% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustifyR-package.R
\docType{package}
\name{clustifyr-package}
\alias{clustifyr-package}
\alias{_PACKAGE}
\title{clustifyr: Classifier for Single-cell RNA-seq Using Cell Clusters}
\description{
Package designed to aid in classifying cells from single-cell RNA sequencing data using external reference data (e.g., bulk RNA-seq, scRNA-seq, microarray, gene lists). A variety of correlation based methods and gene list enrichment methods are provided to assist cell type assignment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rnabioco/clustifyr}
  \item \url{https://rnabioco.github.io/clustifyr/}
  \item Report bugs at \url{https://github.com/rnabioco/clustifyr/issues}
}

}
\author{
\strong{Maintainer}: Kent Riemondy \email{kent.riemondy@cuanschutz.edu}

Authors:
\itemize{
  \item Rui Fu \email{raysinensis@gmail.com}
}

Other contributors:
\itemize{
  \item Austin Gillen \email{austin.gillen@ucdenver.edu} [contributor]
  \item Chengzhe Tian \email{Chengzhe.Tian@colorado.edu} [contributor]
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com} [contributor]
  \item Yue Hao \email{haoyuethink@gmail.com} [contributor]
  \item Michelle Daya \email{michelle.daya@ucdenver.edu} [contributor]
  \item Sidhant Puntambekar \email{sidhantnp@yahoo.com} [contributor]
  \item RNA Bioscience Initiative [funder, copyright holder]
}

}
\keyword{internal}
