% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustifyR.R
\docType{package}
\name{clustifyR}
\alias{clustifyR}
\alias{clustifyR-package}
\alias{clustifyR-package}
\title{clustifyR}
\description{
clustifyR
}
\seealso{
Report bugs at \url{https://github.com/NCBI-Hackathons/clustifyR/issues}
}
\author{
\strong{Maintainer}: Kent Riemondy \email{blah@gmail.com}

Authors:
\itemize{
  \item Austin Gillen \email{blah@gmail.com}
  \item Rui Fu \email{blah@gmail.com}
  \item Jay Hesselberth \email{blah@gmail.com}
  \item Yue Hao \email{blah@gmail.com}
  \item Chengzhe Tian \email{blah@gmail.com}
  \item M. Day \email{blah@gmail.com}
}

}
