% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustifyR-package.R
\docType{package}
\name{clustifyr}
\alias{clustifyr}
\alias{clustifyr-package}
\title{clustifyr}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Classify cells from single-cell RNA sequencing data
    using external reference data (e.g., bulk RNA-seq).
}
\seealso{
Report bugs at \url{https://github.com/rnabioco/clustifyr/issues}
}
\author{
\strong{Maintainer}: Rui Fu \email{rui.fu@ucdenver.edu}

Authors:
\itemize{
  \item Kent Riemondy \email{kent.riemondy@ucdenver.edu}
  \item Austin Gillen \email{austin.gillen@ucdenver.edu}
  \item Chengzhe Tian \email{Chengzhe.Tian@colorado.edu}
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com}
}

Other contributors:
\itemize{
  \item Yue Hao \email{haoyuethink@gmail.com} [contributor]
  \item Michelle Daya \email{michelle.daya@ucdenver.edu} [contributor]
}

}
