% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genelist.R
\name{compare_lists}
\alias{compare_lists}
\title{calculate adjusted p-values for hypergeometric test of gene lists
or jaccard index}
\usage{
compare_lists(bin_mat, marker_m, n = 30000, metric = "hyper")
}
\arguments{
\item{bin_mat}{binarized single-cell expression matrix}

\item{marker_m}{matrix or dataframe of candidate genes for each cluster}

\item{n}{number of genes in the genome}

\item{metric}{adjusted p-value for hypergeometric test, or jaccard index}
}
\description{
calculate adjusted p-values for hypergeometric test of gene lists
or jaccard index
}
