% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_mean_expr}
\alias{compute_mean_expr}
\title{compute mean of clusters}
\usage{
compute_mean_expr(expr_mat, sc_assign, sc_clust)

compute_mean_expr(expr_mat, sc_assign, sc_clust)
}
\description{
compute mean of clusters

compute mean of clusters
}
