% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dplyr.R
\name{cor_to_call}
\alias{cor_to_call}
\title{get best calls for each cluster}
\usage{
cor_to_call(correlation_matrix, metadata = NULL,
  cluster_col = "cluster", collapse_to_cluster = FALSE,
  threshold = 0, rename_suff = NULL)
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{metadata}{input metadata with tsne or umap coordinates and cluster ids}

\item{cluster_col}{metadata column, can be cluster or cellid}

\item{collapse_to_cluster}{if a column name is provided, takes the most frequent call of entire cluster to color in plot}

\item{threshold}{minimum correlation coefficent cutoff for calling clusters}

\item{rename_suff}{suffix to add to type and r column names}
}
\value{
dataframe of cluster, new ident, and r info
}
\description{
get best calls for each cluster
}
