% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_to_call}
\alias{cor_to_call}
\title{get best calls for each cluster}
\usage{
cor_to_call(correlation_matrix, metadata = NULL, col = "cluster",
  collapse_to_cluster = FALSE, threshold = 0)
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{metadata}{input metadata with tsne coordinates and cluster ids}

\item{col}{metadata column, can be cluster or cellid}

\item{collapse_to_cluster}{if a column name is provided, takes the most frequent call of entire cluster to color in plot}

\item{threshold}{minimum correlation coefficent cutoff for calling clusters}
}
\description{
get best calls for each cluster
}
