% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_to_call}
\alias{cor_to_call}
\title{get best calls for each cluster}
\usage{
cor_to_call(correlation_matrix, metadata = NULL,
  cluster_col = "cluster", collapse_to_cluster = FALSE,
  threshold = 0, rename_suff = NULL)
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{metadata}{input metadata with tsne or umap coordinates and cluster ids}

\item{cluster_col}{metadata column, can be cluster or cellid}

\item{collapse_to_cluster}{if a column name is provided, takes the most frequent call of entire cluster to color in plot}

\item{threshold}{minimum correlation coefficent cutoff for calling clusters}

\item{rename_suff}{suffix to add to type and r column names}
}
\description{
get best calls for each cluster
}
\examples{
res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = pbmc_bulk_matrix,
  query_genes = pbmc_vargenes,
  cluster_col = "classified"
)

call1 <- cor_to_call(
  res,
  metadata = pbmc_meta,
  cluster_col = "classified",
  collapse_to_cluster = FALSE,
  threshold = 0.5
)
}
