% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{corr_coef}
\alias{corr_coef}
\title{Correlation function}
\usage{
corr_coef(vec1, vec2, method = "pearson")
}
\arguments{
\item{vec1}{test vector}

\item{vec2}{reference vector}

\item{method}{pearson, spearman, cosine}
}
\description{
Compute correlation between two vectors.
Returned value is between -1 and 1, where 1 indicates highest
positive correlation
and -1 indicates highest negative correlation.
}
