% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downrefs}
\alias{downrefs}
\title{table of references stored in clustifyrdata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 6 columns.
}
\source{
various packages
}
\usage{
downrefs
}
\description{
table of references stored in clustifyrdata
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{cbmc_ref}},
\code{\link{human_genes_10x}},
\code{\link{mouse_genes_10x}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}}
}
\concept{data}
\keyword{datasets}
