% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_marker_parse}
\alias{file_marker_parse}
\title{takes files with positive and negative markers, as described in garnett, and returns list of markers}
\usage{
file_marker_parse(filename)
}
\arguments{
\item{filename}{txt file to load}
}
\value{
list of positive and negative gene markers
}
\description{
takes files with positive and negative markers, as described in garnett, and returns list of markers
}
\examples{
marker_file <- system.file(
  "extdata",
  "hsPBMC_markers.txt",
  package = "clustifyr"
)

file_marker_parse(marker_file)
}
