% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_rank_bias}
\alias{find_rank_bias}
\title{Find rank bias}
\usage{
find_rank_bias(mat, metadata, type_col, ref_mat, query_genes = NULL,
  filter_out = TRUE)
}
\arguments{
\item{mat}{original query expression matrix}

\item{metadata}{metadata after calling types}

\item{type_col}{column name in metadata that contains ids}

\item{ref_mat}{reference expression matrix}

\item{query_genes}{original vector of genes used to clustify}

\item{filter_out}{whether to only report filtered results}
}
\value{
matrix of rank diff values
}
\description{
Find rank bias
}
