% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_rank_bias}
\alias{find_rank_bias}
\title{Find rank bias}
\usage{
find_rank_bias(avg_mat, ref_mat, query_genes = NULL)
}
\arguments{
\item{avg_mat}{average expression matrix}

\item{ref_mat}{reference expression matrix}

\item{query_genes}{original vector of genes used to clustify}
}
\value{
list of matrix of rank diff values
}
\description{
Find rank bias
}
\examples{
avg <- average_clusters(
    mat = pbmc_matrix_small,
    metadata = pbmc_meta,
    cluster_col = "classified",
    if_log = FALSE
)

rankdiff <- find_rank_bias(
    avg,
    cbmc_ref,
    query_genes = pbmc_vargenes
)
}
