% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gene_pct}
\alias{gene_pct}
\title{pct of cells in each cluster that express genelist}
\usage{
gene_pct(matrix, genelist, clusters, returning = "mean")
}
\arguments{
\item{matrix}{expression matrix}

\item{genelist}{vector of marker genes for one identity}

\item{clusters}{vector of cluster identities}

\item{returning}{whether to return mean, min, or max of the gene pct in the gene list}
}
\value{
vector of numeric values
}
\description{
pct of cells in each cluster that express genelist
}
\examples{
res <- gene_pct(
  pbmc_matrix_small,
  cbmc_m$B,
  pbmc_meta$classified
)
}
