% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gene_pct_markerm}
\alias{gene_pct_markerm}
\title{pct of cells in every cluster that express a series of genelists}
\usage{
gene_pct_markerm(matrix, marker_m, cluster_info, cluster_col = NULL,
  norm = NULL)
}
\arguments{
\item{matrix}{expression matrix}

\item{marker_m}{matrixized markers}

\item{cluster_info}{data.frame or vector containing cluster assignments per cell.
Order must match column order in supplied matrix. If a data.frame
provide the cluster_col parameters.}

\item{cluster_col}{column in cluster_info with cluster number}

\item{norm}{whether and how the results are normalized}
}
\description{
pct of cells in every cluster that express a series of genelists
}
\examples{
res <- gene_pct_markerm(pbmc_matrix_small,
                         cbmc_m,
                         pbmc_meta,
                         cluster_col = "classified"
)
}
