% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_best_match_matrix}
\alias{get_best_match_matrix}
\title{Function to make best call from correlation matrix}
\usage{
get_best_match_matrix(cor_mat)
}
\arguments{
\item{cor_mat}{correlation matrix}
}
\value{
matrix of 1s and 0s
}
\description{
Function to make best call from correlation matrix
}
