% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_best_match_matrix}
\alias{get_best_match_matrix}
\title{Function to make best call from correlation matrix}
\usage{
get_best_match_matrix(cor_mat)
}
\arguments{
\item{cor_mat}{correlation matrix}
}
\description{
Function to make best call from correlation matrix
}
\examples{
cor_mat <- res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  cluster_col = "classified",
  ref_mat = pbmc_bulk_matrix,
  query_genes = pbmc_vargenes
)

best_mat <- get_best_match_matrix(cor_mat)
}
