% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_common_elements}
\alias{get_common_elements}
\title{Find entries shared in all vectors}
\usage{
get_common_elements(...)
}
\arguments{
\item{...}{vectors}
}
\value{
vector of shared elements
}
\description{
return entries found in all supplied vectors.
If the vector supplied is NULL or NA, then it will be excluded
from the comparison.
}
