% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbrowsers.R
\name{get_ucsc_reference}
\alias{get_ucsc_reference}
\title{Build reference atlases from external UCSC cellbrowsers}
\usage{
get_ucsc_reference(cb_url, cluster_col, ...)
}
\arguments{
\item{cb_url}{URL of cellbrowser dataset (e.g. http://cells.ucsc.edu/?ds=cortex-dev).
Note that the URL must contain the ds=dataset-name suffix.}

\item{cluster_col}{annotation field for summarizing gene expression (e.g. clustering,
cell-type name, samples, etc.)}

\item{...}{additional args passed to average_clusters}
}
\description{
Build reference atlases from external UCSC cellbrowsers
}
\examples{

\donttest{
get_ext_reference(cb_url = "http://cells.ucsc.edu/?ds=kidney-atlas\%2FFetal_Immune",
                 cluster_col = "celltype")
}
}
