% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genelist.R
\name{get_vargenes}
\alias{get_vargenes}
\title{generate variable gene list from marker matrix}
\usage{
get_vargenes(marker_m)
}
\arguments{
\item{marker_m}{matrix or dataframe of candidate genes for each cluster}
}
\description{
Variable gene list is required for \code{clustify} main function. This
function parses variables genes from a matrix input.
}
