% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{install_clustifyr_app}
\alias{install_clustifyr_app}
\title{Install all packages needed for shiny app}
\usage{
install_clustifyr_app()
}
\value{
installation complete or error messages
}
\description{
Install all packages needed for shiny app
}
\examples{
\dontrun{
install_clustifyr_app()
}
}
