% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{kl_divergence}
\alias{kl_divergence}
\title{KL divergence}
\usage{
kl_divergence(vec1, vec2, if_logcounts = FALSE, total_reads = 1000,
  max_KL = 1)
}
\arguments{
\item{vec1}{test vector}

\item{vec2}{reference vector}

\item{if_logcounts}{pearson, spearman, cosine}

\item{total_reads}{library size}

\item{max_KL}{max_KL}
}
\description{
use package entropy to compute kl divergence
}
