% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_comb_ref}
\alias{make_comb_ref}
\title{make combination ref matrix to assess intermixing}
\usage{
make_comb_ref(ref_mat, if_log = TRUE, sep = "_and_")
}
\arguments{
\item{ref_mat}{reference expression matrix}

\item{if_log}{whether input data is natural}

\item{sep}{separator for name combinations}
}
\value{
expression matrix
}
\description{
make combination ref matrix to assess intermixing
}
\examples{
ref2 <- make_comb_ref(
  cbmc_ref,
  sep = "_+_"
)
}
