% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genelist.R
\name{matrixize_markers}
\alias{matrixize_markers}
\title{convert candidate genes list into matrix}
\usage{
matrixize_markers(marker_df, ranked = FALSE, weight = 0)
}
\arguments{
\item{marker_df}{dataframe of candidate genes}

\item{ranked}{unranked gene list feeds into hyperp, theranked gene list feeds into regular corr_coef}

\item{weight}{ranked genes are tranformed into pseudo expression with added weight}

\item{marker_df}{dataframe of candidate genes}
}
\description{
convert candidate genes list into matrix
}
