% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{object_loc_lookup}
\alias{object_loc_lookup}
\title{lookup table for single cell object structures}
\format{XXX}
\source{
various packages
}
\usage{
object_loc_lookup
}
\description{
lookup table for single cell object structures
}
\keyword{datasets}
