% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_loc_object}
\alias{parse_loc_object}
\title{more flexible parsing of single cell objects}
\usage{
parse_loc_object(
  input,
  type = class(input),
  expr_loc = NULL,
  meta_loc = NULL,
  var_loc = NULL,
  cluster_col = NULL,
  lookuptable = NULL
)
}
\arguments{
\item{input}{input object}

\item{type}{look up predefined slots/loc}

\item{expr_loc}{expression matrix location}

\item{meta_loc}{metadata location}

\item{var_loc}{variable genes location}

\item{cluster_col}{column of clustering from metadata}

\item{lookuptable}{if not supplied, will look in built-in table for object parsing}
}
\value{
list of expression, metadata, vargenes, cluster_col info from object
}
\description{
more flexible parsing of single cell objects
}
\examples{
parse_loc_object(s_small3)
}
