% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_matrix}
\alias{pbmc4k_matrix}
\title{Matrix of single-cell RNA-seq PBMCs.}
\format{A sparseMatrix with genes as rows and cells as columns.

\describe{
}}
\source{
\link{google.com}
}
\usage{
pbmc4k_matrix
}
\description{
A sample of 4000 PBMC cells from a bigger experiment
}
\keyword{datasets}
