% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_matrix}
\alias{pbmc4k_matrix}
\title{Matrix of single-cell RNA-seq PBMCs.}
\format{A sparseMatrix with genes as rows and cells as columns.

\describe{
}
R}
\source{
(https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k)
}
\usage{
pbmc4k_matrix
}
\description{
A sample of 300 PBMC cells, downsampled from a bigger experiment, 30 cells per cluster
}
\keyword{datasets}
