% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc4k_meta}
\alias{pbmc4k_meta}
\title{Meta-data for single-cell RNA-seq PBMCs.}
\format{A data frame with 13 variables:
\describe{
}}
\source{
\url{google.com}
}
\usage{
pbmc4k_meta
}
\description{
A sample of 4000 cells from a bigger experiment
}
\keyword{datasets}
