% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers_M3Drop}
\alias{pbmc_markers_M3Drop}
\title{Marker genes identified by M3Drop from single-cell RNA-seq PBMCs.}
\format{
A data frame with 3 variables:
}
\source{
\verb{[pbmc_matrix]} processed by \verb{[M3Drop]}
}
\usage{
pbmc_markers_M3Drop
}
\description{
Selected features of 3k pbmcs from Seurat3 tutorial
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{cbmc_ref}},
\code{\link{downrefs}},
\code{\link{human_genes_10x}},
\code{\link{mouse_genes_10x}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}}
}
\concept{data}
\keyword{datasets}
