% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_pca}
\alias{pbmc_pca}
\title{Principal components analysis of XXX}
\format{XXX
\describe{
}}
\source{
\link{google.com}
}
\usage{
pbmc_pca
}
\description{
Principal components analysis of XXX
}
\keyword{datasets}
