% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_vargenes}
\alias{pbmc_vargenes}
\title{Variable genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{character} of length 2000.
}
\source{
\verb{[pbmc_matrix]} processed by Seurat
}
\usage{
pbmc_vargenes
}
\description{
Top 2000 variable genes from 3k pbmcs from Seurat3 tutorial
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{cbmc_ref}},
\code{\link{downrefs}},
\code{\link{human_genes_10x}},
\code{\link{mouse_genes_10x}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}}
}
\concept{data}
\keyword{datasets}
