% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent_clusters}
\alias{percent_clusters}
\title{Percentage detected per cluster}
\usage{
percent_clusters(mat, cluster_info, cluster_col = "cluster",
  cut_num = 0.5)
}
\arguments{
\item{mat}{expression matrix}

\item{cluster_info}{data.frame with cells}

\item{cluster_col}{column in cluster_info with cluster number}

\item{cut_num}{binary cutoff for detection}
}
\value{
matrix of numeric values, with genes for row names, and clusters for column names
}
\description{
Percentage detected per cluster
}
