% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_similarity.R
\name{permutation_similarity}
\alias{permutation_similarity}
\title{Compute the p-value for similarity using permutation}
\usage{
permutation_similarity(sc_expr, bulk_expr, sc_meta, num_perm, per_cell = F,
  compute_method, metadata = "cluster", ...)
}
\arguments{
\item{sc_expr}{single-cell expression matrix}

\item{bulk_expr}{bulk expression matrix}

\item{sc_meta}{clustering info of single-cell data assume that genes have ALREADY BEEN filtered}

\item{num_perm}{number of permutations}

\item{per_cell}{run per cell?}

\item{compute_method}{method(s) for computing similarity scores}

\item{metadata}{metadata column used for clustering}

\item{...}{additional paramters to pass to run_one_round}
}
\description{
Permute cluster labels to calculate empirical p-value
}
