% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_similarity.R
\name{permutation_similarity}
\alias{permutation_similarity}
\title{Compute the p-value for similarity using permutation}
\usage{
permutation_similarity(expr_mat, metadata, bulk_mat, num_perm, per_cell = F,
  compute_method, cluster_col = "cluster", ...)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{metadata}{clustering info of single-cell data assume that
genes have ALREADY BEEN filtered}

\item{bulk_mat}{bulk expression matrix}

\item{num_perm}{number of permutations}

\item{per_cell}{run per cell?}

\item{compute_method}{method(s) for computing similarity scores}

\item{cluster_col}{cluster_col column used for clustering}

\item{...}{additional paramters to pass to run_one_round}
}
\description{
Permute cluster labels to calculate empirical p-value
}
