% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{permute_similarity}
\alias{permute_similarity}
\title{Compute a p-value for similarity using permutation}
\usage{
permute_similarity(
  expr_mat,
  ref_mat,
  cluster_ids,
  n_perm,
  per_cell = FALSE,
  compute_method,
  pseudobulk_method = "mean",
  rm0 = FALSE,
  ...
)
}
\arguments{
\item{expr_mat}{single-cell expression matrix}

\item{ref_mat}{reference expression matrix}

\item{cluster_ids}{clustering info of single-cell data assume that
genes have ALREADY BEEN filtered}

\item{n_perm}{number of permutations}

\item{per_cell}{run per cell?}

\item{compute_method}{method(s) for computing similarity scores}

\item{pseudobulk_method}{method used for summarizing clusters, options are mean (default), median, truncate (10\% truncated mean), or trimean, max, min}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{...}{additional parameters}
}
\value{
matrix of numeric values
}
\description{
Permute cluster labels to calculate empirical p-value
}
