% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_best_call}
\alias{plot_best_call}
\title{Plot best calls for each cluster on a tSNE}
\usage{
plot_best_call(correlation_matrix, metadata, col = "cluster",
  collapse_to_cluster = FALSE, threshold = 0, x = "tSNE_1",
  y = "tSNE_2", plot_r = FALSE, ...)
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{metadata}{input metadata with tsne coordinates and cluster ids}

\item{col}{metadata column, can be cluster or cellid}

\item{collapse_to_cluster}{if a column name is provided, takes the most frequent call of entire cluster to color in plot}

\item{threshold}{minimum correlation coefficent cutoff for calling clusters}

\item{x}{x variable}

\item{y}{y variable}

\item{plot_r}{whether to include second plot of cor eff for best call}

\item{...}{passed to plot_tsne}
}
\description{
Plot best calls for each cluster on a tSNE
}
