% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot similarity measures on a tSNE}
\usage{
plot_cor(correlation_matrix, meta_data,
  bulk_data_to_plot = colnames(correlation_matrix),
  metadata.col = "cluster")
}
\arguments{
\item{correlation_matrix}{input similarity matrix}

\item{meta_data}{input metadata with tsne coordinates and cluster ids}

\item{bulk_data_to_plot}{colname of data to plot, defaults to all}

\item{...}{passed to plot_tsne}
}
\description{
Plot similarity measures on a tSNE
}
