% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot a tSNE colored by feature.}
\usage{
plot_tsne(data, x = "tSNE_1", y = "tSNE_2", feature, legend_name = "",
  cols = pretty_palette, pt_size = 0.25)
}
\arguments{
\item{data}{input data}

\item{x}{x variable}

\item{y}{y variable}

\item{feature}{feature to color by}
}
\description{
Plot a tSNE colored by feature.
}
