% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_matrix.R
\name{reduce_matrix}
\alias{reduce_matrix}
\title{Reduce matrix}
\usage{
reduce_matrix(mat, meta, vargene)
}
\arguments{
\item{mat}{expression matrix with row names as the gene names and each column for different cell}

\item{meta}{cluster and cell type info for each cell}

\item{vargene}{a list of gene names for highly variable genes}
}
\description{
Reduce the full expression matrix to only the highly variable genes, change the readcounts to binary factors,
prepare the input data for random forest
}
