% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_clusters.R
\name{refine_clusters}
\alias{refine_clusters}
\title{Refine clustering using reference data}
\usage{
refine_clusters(expr_mat, sc_cluster, bulk_mat, lambda = 0.5, epsilon = 1,
  if_compute_sigma = TRUE, num_iteration = 100, default_similiarity = -5,
  disagreement_freq = 0.001, if_plot = FALSE, tsne_coord = NULL,
  cluster_names = NULL, compute_method, ...)
}
\arguments{
\item{expr_mat}{input single cell data after gene subsetting}

\item{sc_cluster}{initial clustering assignment.
Note that the number of bulk cell types must be no more than
the number of clusters. for additional clusters,
there is an assignment of default_similiarity}

\item{bulk_mat}{input bulk data after gene subsetting}

\item{lambda}{control parameters}

\item{epsilon}{control parameters}

\item{if_compute_sigma}{whether compute the standard deviation
of each gene's expression within each cluster.
If not, the function assumes that all standard deviations equal to 1.}

\item{num_iteration}{maximal number of iteration before exiting
the algorithm}

\item{default_similiarity}{default similarity score if there is
no reference data to compare to}

\item{disagreement_freq}{maximal fraction of changes in cluster
assignment after one iteration to be considered as converged.}

\item{if_plot}{whether to plot the clustering result at each iteration.}

\item{tsne_coord}{t-SNE coordinate of every cells for plotting}

\item{compute_method}{corr_coef}

\item{...}{control parameters for calculating similarity score}

\item{cluster_name}{Name of each cluster for plotting}
}
\description{
Refine clustering using reference data
}
