% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_background}
\alias{remove_background}
\title{Remove high background expression genes from matrix}
\usage{
remove_background(mat, background, n = 0)
}
\arguments{
\item{mat}{expression matrix}

\item{background}{vector or dataframe or matrix of high expression genes in background}

\item{n}{the number of top genes to exclude, 0 defaults to all}
}
\description{
Remove high background expression genes from matrix
}
