% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_background}
\alias{remove_background}
\title{Remove high background expression genes from matrix}
\usage{
remove_background(mat, background, n = 0)
}
\arguments{
\item{mat}{expression matrix}

\item{background}{vector or dataframe or matrix of high expression genes in background}

\item{n}{the number of top genes to exclude, 0 defaults to all}
}
\value{
expression matrix with rows removed
}
\description{
Remove high background expression genes from matrix
}
\examples{
avg1 <- average_clusters_filter(
  mat = pbmc_matrix_small,
  metadata = pbmc_meta,
  filter_on = "nFeature_RNA"
)

remove_background(pbmc_matrix_small, avg1, 1)
}
