% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowVar}
\alias{RowVar}
\title{get var per row for matrix}
\usage{
RowVar(x, na.rm = T)
}
\arguments{
\item{x}{expression matrix}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations?}
}
\description{
get var per row for matrix
}
