% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_clustifyr_app}
\alias{run_clustifyr_app}
\title{Launch Shiny app version of clustifyr}
\usage{
run_clustifyr_app()
}
\value{
instance of shiny app
\donttest{
run_clustifyr_app()
}
}
\description{
Launch Shiny app version of clustifyr
}
