% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_cor}
\alias{run_cor}
\title{run_cor: Main function to compare scRNA-seq data to bulk RNA-seq data.}
\usage{
run_cor(sc_expr, sc_meta, bulk_expr, query_gene_list, per_cell = F,
  if_permute = TRUE, num_permute = 1000, return_full = F, compute_method,
  ...)
}
\arguments{
\item{sc_expr}{single-cell expression matrix}

\item{sc_meta}{clustering info of single-cell data assume that genes have ALREADY BEEN filtered}

\item{bulk_expr}{bulk expression matrix}

\item{query_gene_list}{A vector of genes of interest.}

\item{per_cell}{run per cell?}

\item{compute_method}{method(s) for computing similarity scores}

\item{...}{additional paramters to pass to run_one_round}
}
\description{
run_cor: Main function to compare scRNA-seq data to bulk RNA-seq data.
}
