% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_small}
\alias{s_small}
\title{Small clustered Seurat2 object}
\format{XXX
\describe{
}}
\source{
\code{[pbmc_small]} processed by seurat
}
\usage{
s_small
}
\description{
Small clustered Seurat2 object
}
\keyword{datasets}
