% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_small}
\alias{s_small}
\title{Small clustered Seurat object}
\format{XXX}
\source{
\code{[pbmc_small]} processed by \code{[seurat]}
}
\usage{
s_small
}
\description{
Small clustered Seurat object
}
\keyword{datasets}
