% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_small3}
\alias{s_small3}
\title{Small clustered Seurat3 object}
\format{XXX
\describe{
}}
\source{
\code{[pbmc_small]} processed by \code{[Seurat]}
}
\usage{
s_small3
}
\description{
Small clustered Seurat3 object
}
\keyword{datasets}
