% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_small3}
\alias{s_small3}
\title{Small clustered Seurat3 object}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\source{
\verb{[pbmc_small]} processed by Seurat
}
\usage{
s_small3
}
\description{
Small clustered Seurat3 object
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{cbmc_ref}},
\code{\link{downrefs}},
\code{\link{human_genes_10x}},
\code{\link{mouse_genes_10x}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}},
\code{\link{s_small}},
\code{\link{sce_small}}
}
\concept{data}
\keyword{datasets}
