% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_small}
\alias{sce_small}
\title{Small SingleCellExperiment object}
\format{
An object of class \code{SingleCellExperiment} with 200 rows and 200 columns.
}
\source{
\url{https://github.com/hemberg-lab/scRNA.seq.datasets/blob/master/R/segerstolpe.R}
}
\usage{
sce_small
}
\description{
Small SingleCellExperiment object
}
\seealso{
Other data: 
\code{\link{cbmc_m}},
\code{\link{cbmc_ref}},
\code{\link{downrefs}},
\code{\link{human_genes_10x}},
\code{\link{mouse_genes_10x}},
\code{\link{object_loc_lookup}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix_small}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}},
\code{\link{s_small3}},
\code{\link{s_small}}
}
\concept{data}
\keyword{datasets}
