% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{seurat_meta}
\alias{seurat_meta}
\alias{seurat_meta.seurat}
\alias{seurat_meta.Seurat}
\title{Function to convert labelled seurat object to fully prepared metadata#'}
\usage{
seurat_meta(seurat_object, ...)

\method{seurat_meta}{seurat}(seurat_object, dr = "umap", ...)

\method{seurat_meta}{Seurat}(seurat_object, dr = "umap", ...)
}
\arguments{
\item{seurat_object}{seurat_object after tsne or umap projections and clustering}

\item{...}{additional arguments}

\item{dr}{dimension reduction method}
}
\value{
dataframe of metadata, including dimension reduction plotting info
}
\description{
Function to convert labelled seurat object to fully prepared metadata#'
}
\examples{
\dontrun{
meta_data <- seurat_meta(s_small)
}
}
