% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{seurat_ref}
\alias{seurat_ref}
\alias{seurat_ref.seurat}
\alias{seurat_ref.Seurat}
\title{Function to convert labelled seurat object to avg expression matrix}
\usage{
seurat_ref(seurat_object, ...)

\method{seurat_ref}{seurat}(seurat_object, cluster_col = "classified",
  var_genes_only = FALSE, assay_name = NULL, method = "mean", ...)

\method{seurat_ref}{Seurat}(seurat_object, cluster_col = "classified",
  var_genes_only = FALSE, assay_name = NULL, method = "mean", ...)
}
\arguments{
\item{seurat_object}{seurat_object after tsne or umap projections and clustering}

\item{...}{additional arguments}

\item{cluster_col}{column name where classified cluster names are stored in seurat meta data, cannot be "rn"}

\item{var_genes_only}{whether to keep only var_genes in the final matrix output, could also look up genes used for PCA}

\item{assay_name}{any additional assay data, such as ADT, to include. If more than 1, pass a vector of names}

\item{method}{whether to take mean (default) or median}

\item{seurat_object}{seurat_object after tsne or umap projections and clustering}

\item{cluster_col}{column name where classified cluster names are stored in seurat meta data, cannot be "rn"}

\item{var_genes_only}{whether to keep only var_genes in the final matrix output, could also look up genes used for PCA}

\item{assay_name}{any additional assay data, such as ADT, to include. If more than 1, pass a vector of names}

\item{method}{whether to take mean (default) or median}

\item{...}{additional arguments}
}
\value{
reference expression matrix, with genes as row names, and cell types as column names

reference expression matrix

reference expression matrix
}
\description{
Function to convert labelled seurat object to avg expression matrix
}
\examples{
avg <- seurat_ref(
  seurat_object = s_small,
  cluster_col = "res.1",
  var_genes_only = TRUE
)
}
