% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrapper.R
\name{write_meta}
\alias{write_meta}
\alias{write_meta.seurat}
\alias{write_meta.Seurat}
\title{Function to write metadata to object}
\usage{
write_meta(object, ...)

\method{write_meta}{seurat}(object, meta)

\method{write_meta}{Seurat}(object, meta)
}
\arguments{
\item{object}{object after tsne or umap projections
and clustering}

\item{meta}{new metadata dataframe}
}
\value{
object with newly inserted metadata columns
}
\description{
Function to write metadata to object
}
\examples{
obj <- write_meta(
    object = s_small,
    meta = seurat_meta(s_small)
)
ref[1:3, 1:3]
obj <- write_meta(
    object = s_small,
    meta = seurat_meta(s_small)
)
ref[1:3, 1:3]
}
